+++ packages/node/src/indexer/dictionary.service.ts
@@ -3,6 +3,7 @@
 
 import { Inject, Injectable } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
+import { NETWORK_FAMILY } from '@subql/common';
 import {
   NodeConfig,
   DictionaryService as CoreDictionaryService,
@@ -11,17 +12,34 @@ import { SubqueryProject } from '../configure/SubqueryProject';
 
 @Injectable()
 export class DictionaryService extends CoreDictionaryService {
-  constructor(
+  private constructor(
     @Inject('ISubqueryProject') protected project: SubqueryProject,
     nodeConfig: NodeConfig,
     eventEmitter: EventEmitter2,
+    dictionaryUrl?: string,
   ) {
     super(
-      project.network.dictionary,
+      dictionaryUrl ?? project.network.dictionary,
       project.network.chainId,
       nodeConfig,
       eventEmitter,
       ['lastProcessedHeight', 'chain'],
     );
+  }
+
+  static async create(
+    project: SubqueryProject,
+    nodeConfig: NodeConfig,
+    eventEmitter: EventEmitter2,
+  ): Promise<DictionaryService> {
+    const url =
+      project.network.dictionary ??
+      (await CoreDictionaryService.resolveDictionary(
+        NETWORK_FAMILY.cosmos,
+        project.network.chainId,
+        nodeConfig.dictionaryRegistry,
+      ));
+
+    return new DictionaryService(project, nodeConfig, eventEmitter, url);
   }
 }