+++ packages/node/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Updates to match changes in `@subql/node-core` (#209)
+  - Dictionary service to use dictionary registry
+  - Use yargs from node core
 
 ## [3.3.1] - 2023-11-08
 ### Fixed